requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.37.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
Pillow==10.4.0
openai==1.51.2
fugashi==1.3.2
unidic-lite==1.0.8
rank-bm25==0.2.2
streamlit-geolocation==0.0.10
beautifulsoup4==4.12.3
lxml==5.3.0
pdfminer.six==20231228

